requirement.txt
playwright==1.54.0
pytest==8.3.2
pytest-playwright==0.5.2
python-dotenv==1.0.1